% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{aliases}
\alias{aliases}
\alias{AND}
\alias{OR}
\alias{nor}
\alias{neither}
\alias{NOR}
\alias{NEITHER}
\alias{pow}
\alias{XOR}
\title{Aliases}
\usage{
AND(x, y)

OR(x, y)

nor(x, y)

neither(x, y)

NOR(x, y)

NEITHER(x, y)

pow()

XOR(x, y)
}
\arguments{
\item{x, y}{Logical conditions.}
}
\description{
These simple aliases can be useful to avoid operator precedence ambiguity,
or to make use of indents from commas within your text editor. The all-caps versions accept
single-length (capable of 'short-circuits') logical conditions only.

Neithers and nors are identical except have slightly different short-circuits.
 \code{NOR} uses negation once so may be quicker if the first argument is very, very prompt.
}
