% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datakable.R
\name{data.kable}
\alias{data.kable}
\title{Print data like print.data.table}
\usage{
data.kable(DT, ...)
}
\arguments{
\item{DT}{A \code{data.frame} or matrix coercible to such.}

\item{...}{Arguments passed to \code{\link[knitr]{kable}}. (\code{format.args} is used here.)}
}
\description{
I like the \code{print} method of \code{data.table}s. This is a simple
function to replicate it using \code{\link[knitr]{kable}}.
}
