% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_copy2a4.R
\name{dev_copy2a4}
\alias{dev_copy2a4}
\title{Copy device to an A4 PDF}
\usage{
dev_copy2a4(filename, landscape = TRUE, ...)
}
\arguments{
\item{filename}{A string giving the name of the PDF file to write to, must end in \code{.pdf}.}

\item{...}{Other parameters passed to \code{\link[grDevices]{pdf}}.}
}
\value{
As in \code{\link[grDevices]{dev2}}.
}
\description{
Simply a wrapper around \code{dev.copy2pdf}, but without the need to remember that an A4 sheet of paper is 8.27 in by 11.69 in.
}
