% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir2.R
\name{dir2}
\alias{dir2}
\title{List many files}
\usage{
dir2(
  path = ".",
  file_ext = NULL,
  full.names = TRUE,
  recursive = TRUE,
  pattern = NULL,
  fixed = FALSE,
  perl = TRUE && missing(fixed) && !fixed,
  ignore.case = FALSE,
  invert = FALSE,
  .dont_use = FALSE
)
}
\arguments{
\item{path}{A string representing the trunk path to search within.}

\item{file_ext}{A string like `*.txt` or `.csv` to limit the result to files 
with that extension.}

\item{full.names}{\code{TRUE} by default.}

\item{recursive}{\code{TRUE} by default.}

\item{pattern, perl, ignore.case, fixed, invert}{As in \code{grep} but with different defaults.
Used to filter files with extension \code{file_ext}.}

\item{.dont_use}{Only used for tests to simulate non-Windows systems.}
}
\value{
The same as \code{\link[base]{list.files}}, a character vector of files sought.
}
\description{
(Windows only) Same as \code{\link[base]{list.files}} but much faster.

Present since \code{v1.4.0.}
}
