% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_colr.R
\name{drop_colr}
\alias{drop_colr}
\title{Drop columns whose names match a pattern}
\usage{
drop_colr(DT, pattern, ..., checkDT = TRUE)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{pattern}{A regular expression as in \code{grepl}.}

\item{...}{Arguments passed to \code{grepl}.}

\item{checkDT}{If \code{TRUE} (the default), will error if \code{DT} is not a \code{data.table}.}
}
\description{
Drop columns whose names match a pattern
}
