% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_empty_cols.R
\name{drop_empty_cols}
\alias{drop_empty_cols}
\title{Drop empty columns}
\usage{
drop_empty_cols(DT, copy = FALSE)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{copy}{Copies the \code{data.table} so the original can be retained. 
Not applicable if \code{DT} is not a \code{data.table}.
If \code{FALSE}, the default, \code{DT} itself will be modified.}
}
\description{
Removes columns from a \code{data.table} where all the values are missing.
}
