% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_rows.R
\name{duplicated_rows}
\alias{duplicated_rows}
\title{Return duplicated rows of data.table}
\usage{
duplicated_rows(DT, by = names(DT), na.rm = FALSE, order = TRUE)
}
\arguments{
\item{DT}{A data.table.}

\item{by}{Columns to evaluate duplicates over.}

\item{na.rm}{Remove NA is in \code{by} before returning duplicates.}

\item{order}{Order the result so that it duplicates are adjacent.}
}
\value{
Duplicate rows of DT by \code{by}. For interactive use.
}
\description{
Return duplicated rows of data.table
}
