% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notchin.R
\name{\%notchin\%}
\alias{\%notchin\%}
\title{Negation of in (character)}
\usage{
x \%notchin\% y
}
\arguments{
\item{x}{Values to be matched.}

\item{y}{Values to be matched against.}
}
\description{
Negation of in (character)
}
\details{
If \code{y} is \code{NULL}, then \code{x} is \code{TRUE} for consistency with
 \code{\%in\%}. If \code{x} and \code{y} are not both character, the function simply
 falls back to \code{\%in\%} rather than erroring.
}
