% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negation of in}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{Values to be matched}

\item{y}{Values to be matched against.}
}
\description{
Negation of in
}
\details{
If \code{y} is \code{NULL}, then \code{x} is \code{TRUE} for consistency with
 \code{\%in\%}. Note that the function uses \code{\link[fastmatch]{fmatch}} internally for 
performance on large \code{y}. Accordingly, \code{y} will be modified by adding
a \code{.match.hash} attribute and thus must not be used in packages where \code{y}
is a constant, or for things like names of \code{data.table}.
}
