% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{\%pin\%}
\alias{\%pin\%}
\title{Partial in}
\usage{
x \%pin\% Y
}
\arguments{
\item{x}{The values to be matched. Same as \code{\%in\%}.}

\item{Y}{A vector of values (perl regular expressions) to be matched against.}
}
\value{
\code{TRUE} for every \code{x} for which any \code{grepl} is \code{TRUE}.
}
\description{
Analog of \code{\%in\%} but indicating partial match of the left operand.
}
\examples{
x <- c("Sydney Airport", "Melbourne Airport")

x \%pin\% c("Syd", "Melb")



}
