% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haversine_distance.R
\name{haversine_distance}
\alias{haversine_distance}
\title{Distance between two points on the Earth}
\usage{
haversine_distance(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1, lon1, lat2, lon2}{That latitudes and longitudes of the two points.}
}
\value{
The distance in kilometres between the two points.
}
\description{
Distance between two points on the Earth
}
\details{
This is reasonably accurate for distances in the order of 1 to 1000 km.
}
\examples{

# Distance from YMEL to YSSY
haversine_distance(-37 - 40/60, 144 + 50/60, -33 - 56/60, 151 + 10/60)

}
