% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutils-package.R
\name{hutils-package}
\alias{hutils-package}
\title{hutils package}
\description{
Provides utility functions for, and drawing on, the 'data.table' package. The package also collates useful miscellaneous functions extending base R not available elsewhere. The name is a portmanteau of 'utils' and the author.
}
