% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAttached.R
\name{isAttached}
\alias{isAttached}
\title{Is a package attached?}
\usage{
isAttached(pkg)
}
\arguments{
\item{pkg}{Either character or unquoted.}
}
\value{
\code{TRUE} if \code{pkg} is attached.
}
\description{
Is a package attached?
}
