% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTrueFalse.R
\name{isTrueFalse}
\alias{isTrueFalse}
\title{Logical assertions}
\usage{
isTrueFalse(x)
}
\arguments{
\item{x}{An object whose values are to be checked.}
}
\value{
For \code{isTrueFalse}, \code{TRUE} if and only if \code{x} is \code{TRUE} or \code{FALSE} identically (perhaps with attributes).
}
\description{
Logical assertions
}
