% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_na.R
\name{mean_na}
\alias{mean_na}
\title{Proportion of values that are NA.}
\usage{
mean_na(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
A double, \code{mean(is.na(v))}.
}
\description{
Proportion of values that are NA.
}
