% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Statistical mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{A vector for which the mode is desired.}
}
\value{
The most common element of \code{x}.

If the mode is not unique, only one of these values is returned, for simplicity.

If \code{x} has length zero, \code{Mode(x) = x}.
}
\description{
Present since \code{hutils 1.4.0}. The most common element.
}
