% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_files.R
\name{number_of_files}
\alias{number_of_files}
\title{Quickly count the number of files in a directory}
\usage{
number_of_files(dir, filetype = "*.*")
}
\description{
Quickly count the number of files in a directory
}
