% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_transpose_data.table.R
\name{print_transpose_data.table}
\alias{print_transpose_data.table}
\title{Print transposed data.table}
\usage{
print_transpose_data.table(DT, sep = "", file = "", append = FALSE)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{sep}{Passed to \code{\link[base]{cat}}: the separation put horizontally between each row in \code{DT}.}

\item{file}{Passed to \code{\link[base]{cat}}. If \code{file = ""}, the default, 
output is printed to the console. Otherwise, result will be sunk to a file.}

\item{append}{Should the result be appended to \code{file}? Not applicable if \code{file = ""}.}
}
\description{
Print transposed data.table
}
\details{
\code{NA}s will be replaced by \code{.} and backslashes will be replaced by \code{@ } (to avoid backslashes).
}
