% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide-dir.R
\name{provide.dir}
\alias{provide.dir}
\title{Provide directory}
\usage{
provide.dir(path, ...)
}
\arguments{
\item{path}{Path to create.}

\item{...}{Passed to \code{dir.create}.}
}
\description{
Provide directory. Create directory only if it does not exist.
}
