% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_error.R
\name{report_error}
\alias{report_error}
\title{Report errors and warnings}
\usage{
report_error(faulty_input, error_condition, requirement, context = NULL,
  advice, hint = NULL, halt = TRUE)
}
\arguments{
\item{faulty_input}{Unquoted function argument that is the cause of the error condition.}

\item{error_condition}{A sentence explaining the condition that invoked the error.}

\item{requirement}{A sentence that explains what is required.}

\item{context}{(Optional) A sentence that contextualizes the error}

\item{advice}{Advice for the user to avoid the error.}

\item{hint}{If the input can be guessed,}

\item{halt}{(logical, default: \code{TRUE}) Should the function signal an error and halt?}
}
\description{
Provides a consistent style for errors and warnings.
}
