% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQ.R
\name{RQ}
\alias{RQ}
\title{Shorthand for \code{requireNamespace}}
\usage{
RQ(pkg, yes, no)
}
\arguments{
\item{pkg}{Packages to test whether the package is not yet installed.}

\item{yes}{Response if \code{pkg} is \strong{not} installed.}

\item{no}{Response if \code{pkg} is installed.}
}
\description{
\code{if (!requireNamespace(pkg, quietly = TRUE))} \emph{yes} \code{else} \emph{no}.
Typical use-case would be \code{RQ(pkg, install.packages("pkg"))].}

This function is not recommended for use in scripts as it is a bit cryptic; it's 
use-case is for bash scripts and the like where calls like this would otherwise
be frequent and cloud the message.
}
\examples{
\dontrun{
 RQ("dplyr", "dplyr needs installing")
}


}
