% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_grep.R
\name{select_grep}
\alias{select_grep}
\title{Select names matching a pattern}
\usage{
select_grep(DT, patterns, .and = NULL, .but.not = NULL)
}
\arguments{
\item{DT}{A \code{data.frame}.}

\item{patterns}{Regular expressions to be matched against the names of \code{DT}. If \code{length(patterns) > 1} the patterns are concatenated using alternation.}

\item{.and}{Character or integer positions of names to select, regardless of whether or not they are matched by \code{patterns}.}

\item{.but.not}{Character or integer positions of names to drop, regardless of whether or not they are matched by \code{patterns} or whether they are explicitly added by \code{.and}.}
}
\value{
\code{DT} with the selected names.
}
\description{
Select names matching a pattern
}
