% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{selector}
\alias{selector}
\title{Fasting selection of \code{data.table} columns}
\usage{
selector(DT, ..., cols = NULL, preserve.key = TRUE)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{...}{Unquoted columns names.}

\item{cols}{Character vector of column names.}

\item{preserve.key}{Whether to reapply the key, if \code{DT} has one.}
}
\value{
\code{DT} with the selected columns.
}
\description{
Present since \code{hutils 1.2.0}.
}
