% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_nrow.R
\name{seq_nrow}
\alias{seq_nrow}
\title{Generate sequence of row numbers}
\usage{
seq_nrow(x)
}
\arguments{
\item{x}{An object that admits an \code{nrow}.}
}
\value{
Equivalent to \code{seq_len(nrow(x))}
}
\description{
Generate sequence of row numbers
}
