% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cols_first.R
\name{set_cols_first}
\alias{set_cols_first}
\alias{set_cols_last}
\title{Put columns first or last}
\usage{
set_cols_first(DT, cols, intersection = TRUE)

set_cols_last(DT, cols, intersection = TRUE)
}
\arguments{
\item{DT}{A data.table.}

\item{cols}{Columns to put before (after) all others}

\item{intersection}{Use the intersection of the names of \code{DT} and \code{cols}.}
}
\description{
Put columns first or last
}
