% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\alias{\%<->\%}
\title{Swap assignment}
\usage{
x \%<->\% value
}
\arguments{
\item{x, value}{Objects whose values are to be reassigned by swapping.}
}
\value{
\code{NULL} invisibly. Called for its side-effect: the values
of \code{x} and \code{value} are swapped. So
\preformatted{x \%<->\% value} 
is equivalent to
\preformatted{temp <- x
x <- value
value <- temp
rm(temp)}
}
\description{
Swap values simultaneously. Present since \code{hutils 1.4.0}.
}
\examples{
a <- 1
b <- 2
a \%<->\% b
a
b

}
