% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longest_affix.R
\name{trim_common_affixes}
\alias{trim_common_affixes}
\title{Longest common prefix/suffix}
\usage{
trim_common_affixes(x, .x = NULL)
}
\arguments{
\item{x}{A character vector.}

\item{.x}{If \code{NULL}, the default, ignored. May be used if \code{x} is
known to be free of \code{NA}s.}
}
\value{
The longest common substring in \code{x} either at the start or end of each string.
For \code{trim_common_affixes} \code{x} with common prefix and common suffix
removed.
}
\description{
Longest common prefix/suffix
}
\examples{
longest_prefix(c("totalx", "totaly", "totalz"))
longest_suffix(c("ztotal", "ytotal", "xtotal"))
}
