% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueLookup.R
\name{uniqueLookup}
\alias{uniqueLookup}
\title{Pseudo memoize assignment by refernece}
\usage{
uniqueLookup(DT, new_var, old_var, f_old_var, by = NULL)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{new_var}{Name of the new variable.}

\item{old_var}{Variable to be passed to \code{f_old_var}.}

\item{f_old_var}{Function to convert \code{old_var} to \code{new_var}.}
}
\value{
\code{DT} with the update variable.
}
\description{
When a value is repeated often in a large \code{data.table}, it may be more efficient to 
only perform the operation on the unique values.
}
\examples{
dt <- data.table(
  DateChar = as.character(as.Date("2000-01-01") + sample(10, replace = TRUE, size = 1e6))
)
system.time(dt[, Date1 := as.Date(DateChar)])
system.time(uniqueLookup(dt, "Date2", "DateChar", as.Date))
}
