% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_grepl.R
\name{any_grepl}
\alias{any_grepl}
\title{Does the pattern appear anywhere?}
\usage{
any_grepl(x, pattern, perl = TRUE, ignore.case = FALSE,
  fixed = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{A character vector.}

\item{pattern, perl, ignore.case, fixed}{As in \code{\link[base]{grep}}.}

\item{quiet}{(logical, default: \code{FALSE}) If \code{TRUE}, silences any messages.}
}
\description{
Shortcut for \code{any(grepl(...))}, mostly for consistency.
}
\details{
From version \code{v 1.4.0}, \code{any_grepl(a, bb)} will be internally
reversed to \code{any_grepl(bb, a)} if \code{length(bb) > 1} and \code{length(a) == 1}.
}
\examples{
any_grepl(c("A_D_E", "K0j"), "[a-z]")

}
