% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_empty_cols.R
\name{drop_empty_cols}
\alias{drop_empty_cols}
\title{Drop empty columns}
\usage{
drop_empty_cols(DT)
}
\arguments{
\item{DT}{A \code{data.table}.}
}
\description{
Removes columns from a \code{data.table} where all the values are missing.
}
