% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_columns.R
\name{fst_columns}
\alias{fst_columns}
\title{Utilities for `fst` files}
\usage{
fst_columns(file.fst)
}
\arguments{
\item{file.fst}{Path to file.}
}
\value{
Various outputs:
\describe{
\item{\code{fst_columns}}{Returns the names of the columns in \code{file.fst}.}
}
}
\description{
Utilities for `fst` files
}
