% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_LaTeX_manual.R
\name{generate_LaTeX_manual}
\alias{generate_LaTeX_manual}
\title{Generate LaTeX manual of installed package}
\source{
<https://stackoverflow.com/a/30608000/1664978>
}
\usage{
generate_LaTeX_manual(pkg)
}
\arguments{
\item{pkg}{Quoted package name (must be installed).}
}
\value{
See \code{\link[base]{system}}.
Called for its side-effect: creates a PDF in the current working directory.
}
\description{
Generate LaTeX manual of installed package
}
