% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implies.R
\name{implies}
\alias{implies}
\alias{\%implies\%}
\title{Logical implies}
\usage{
implies(x, y)

x \%implies\% y
}
\arguments{
\item{x, y}{Logical vectors of the same length.}
}
\value{
\code{TRUE} along \code{y} unless \code{x} is \code{TRUE} and \code{y} is \code{FALSE}.
}
\description{
Returns the result of \eqn{x\Longrightarrow y}.
}
\examples{

c(TRUE, TRUE, FALSE, FALSE) \%implies\%
c(TRUE, FALSE, TRUE, FALSE)


}
