% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide-file.R
\name{provide.file}
\alias{provide.file}
\title{Provide a file}
\usage{
provide.file(path)
}
\arguments{
\item{path}{A string. The path to a filename that requires existence.}
}
\value{
\code{path} for success. Or the empty string \code{character(1)} if failures.
}
\description{
Provide a file
}
