% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-exchange.R
\name{swap}
\alias{swap}
\alias{\%<->\%}
\title{Swap assignment}
\usage{
x \%<->\% y
}
\arguments{
\item{x, y}{Objects whose values are to be reassigned by swapping.}
}
\value{
\code{NULL} invisibly. Called for its side-effect: the values
of \code{x} and \code{y} are swapped. So
\preformatted{x \%<->\% y} 
is equivalent to
\preformatted{temp <- x
x <- y
y <- temp
rm(temp)}
}
\description{
Swap assignment
}
